+++ test/unit/base.test.js
@@ -144,14 +144,14 @@ describe('CONVERT FUNCTION TESTS ', function() {
           expect(conversionResult.output[0].data).to.have.property('item');
           expect(conversionResult.output[0].data).to.have.property('variable');
           expect(conversionResult.output[0].data.variable).to.be.an('array');
-          expect(conversionResult.output[0].data.variable[1].id).to.equal('format');
+          expect(conversionResult.output[0].data.variable[1].key).to.equal('format');
           expect(conversionResult.output[0].data.variable[1].value).to.equal('json');
-          expect(conversionResult.output[0].data.variable[2].id).to.equal('path');
+          expect(conversionResult.output[0].data.variable[2].key).to.equal('path');
           expect(conversionResult.output[0].data.variable[2].value).to.equal('send-email');
-          expect(conversionResult.output[0].data.variable[3].id).to.equal('new-path-variable-1');
+          expect(conversionResult.output[0].data.variable[3].key).to.equal('new-path-variable-1');
           // serialised value for object { R: 100, G: 200, B: 150 }
           expect(conversionResult.output[0].data.variable[3].value).to.equal('R,100,G,200,B,150');
-          expect(conversionResult.output[0].data.variable[4].id).to.equal('new-path-variable-2');
+          expect(conversionResult.output[0].data.variable[4].key).to.equal('new-path-variable-2');
           // serialised value for array ["exampleString", "exampleString"]
           expect(conversionResult.output[0].data.variable[4].value).to.equal('exampleString,exampleString');
           done();
@@ -406,6 +406,20 @@ describe('CONVERT FUNCTION TESTS ', function() {
               .equal('{\n    "a": "<string>",\n    "b": "<string>"\n}');
             expect(exampleRequest.body.raw).to
               .equal('{\n    "a": "example-b",\n    "b": "example-c"\n}');
+            done();
+          });
+      });
+
+      it('[Github #338] Should contain non-truthy example from examples outside of schema instead of faked value' +
+      examplesOutsideSchema, function(done) {
+        Converter.convert({ type: 'file', data: examplesOutsideSchema },
+          { schemaFaker: true, requestParametersResolution: 'example', exampleParametersResolution: 'example' },
+          (err, conversionResult) => {
+            let rootRequest = conversionResult.output[0].data.item[0].request;
+
+            expect(err).to.be.null;
+            expect(rootRequest.url.query[0].key).to.eql('limit');
+            expect(rootRequest.url.query[0].value).to.eql('0');
             done();
           });
       });
@@ -902,18 +916,19 @@ describe('CONVERT FUNCTION TESTS ', function() {
         requestUrl = conversionResult.output[0].data.item[0].request.url;
         collectionVars = conversionResult.output[0].data.variable;
         expect(requestUrl.host).to.eql(['{{baseUrl}}']);
-        expect(_.find(collectionVars, { id: 'baseUrl' }).value).to.eql('{{BASE_URI}}/api');
-        expect(_.find(collectionVars, { id: 'BASE_URI' }).value).to.eql('https://api.example.com');
+        expect(_.find(collectionVars, { key: 'baseUrl' }).value).to.eql('{{BASE_URI}}/api');
+        expect(_.find(collectionVars, { key: 'BASE_URI' }).value).to.eql('https://api.example.com');
         done();
       });
     });
 
-    it('[Github #31] - should set optional params as disabled', function(done) {
+    it('[Github #31] & [GitHub #337] - should set optional params as disabled', function(done) {
       let options = { schemaFaker: true, disableOptionalParameters: true };
       Converter.convert({ type: 'file', data: requiredInParams }, options, (err, conversionResult) => {
         expect(err).to.be.null;
         let requests = conversionResult.output[0].data.item[0].item,
-          request;
+          request,
+          urlencodedBody;
 
         // GET /pets
         // query1 required, query2 optional
@@ -924,6 +939,13 @@ describe('CONVERT FUNCTION TESTS ', function() {
         expect(request.header[0].disabled).to.be.false;
         expect(request.header[1].disabled).to.be.true;
 
+        // POST /pets
+        // urlencoded body
+        urlencodedBody = requests[2].request.body.urlencoded;
+        expect(urlencodedBody[0].key).to.eql('urlencodedParam1');
+        expect(urlencodedBody[0].disabled).to.be.false;
+        expect(urlencodedBody[1].key).to.eql('urlencodedParam2');
+        expect(urlencodedBody[1].disabled).to.be.true;
         done();
       });
     });