+++ test/unit/util.test.js
@@ -231,7 +231,7 @@ describe('SCHEMA UTILITY FUNCTION TESTS ', function () {
         },
         retVal = SchemaUtils.convertToPmCollectionVariables(serverVariables, null, null);
       expect(retVal).to.be.an('array');
-      expect(retVal[0].id).to.equal('v1');
+      expect(retVal[0].key).to.equal('v1');
       expect(retVal[0].value).to.equal('v2.0');
     });
 
@@ -249,9 +249,9 @@ describe('SCHEMA UTILITY FUNCTION TESTS ', function () {
 
       expect(retVal).to.be.an('array');
 
-      expect(retVal[0].id).to.equal('v1');
+      expect(retVal[0].key).to.equal('v1');
       expect(retVal[0].value).to.equal('v2.0');
-      expect(retVal[1].id).to.equal('baseUrl');
+      expect(retVal[1].key).to.equal('baseUrl');
       expect(retVal[1].value).to.equal('hello.com');
     });
   });
@@ -2016,6 +2016,8 @@ describe('SCHEMA UTILITY FUNCTION TESTS ', function () {
       expect(convertedUrl).to.equal('{{scheme}}://developer.uspto.gov/{{path0}}/segment/{{path1}}');
 
       expect(SchemaUtils.fixPathVariablesInUrl('{{a}}')).to.equal('{{a}}');
+
+      expect(SchemaUtils.fixPathVariablesInUrl('/agents/{agentId}}')).to.equal('/agents/{{agentId}}}');
 
       expect(SchemaUtils.fixPathVariablesInUrl('{{a}}://{b}.com/{pathvar}/{morevar}'))
         .to.equal('{{a}}://{{b}}.com/{{pathvar}}/{{morevar}}');